requirements.txt
asgiref==3.5.0
dj-database-url==0.5.0
Django==4.0.2
django-braces==1.15.0
djangorestframework==3.13.1
gunicorn==20.1.0
Pillow==9.0.1
psycopg2==2.9.3
pytz==2021.3
sqlparse==0.4.2
whitenoise==6.0.0
